+++ examples/SendAndReceive/SendAndReceive.cpp
@@ -1,0 +1,156 @@
+/*
+ * SendAndReceive.cpp
+ *
+ * Demonstrates sending IR codes and receiving it simultaneously
+ *
+ *  This file is part of Arduino-IRremote https://github.com/Arduino-IRremote/Arduino-IRremote.
+ *
+ ************************************************************************************
+ * MIT License
+ *
+ * Copyright (c) 2021 Armin Joachimsmeyer
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is furnished
+ * to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
+ * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE
+ * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ *
+ ************************************************************************************
+ */
+
+#include <Arduino.h>
+
+/*
+ * Define macros for input and output pin etc.
+ */
+#include "PinDefinitionsAndMore.h"
+
+//#define EXCLUDE_EXOTIC_PROTOCOLS // saves around 240 bytes program space if IrSender.write is used
+//#define SEND_PWM_BY_TIMER
+//#define USE_NO_SEND_PWM
+
+#include <IRremote.h>
+
+#define DELAY_AFTER_SEND 2000
+#define DELAY_AFTER_LOOP 5000
+
+void setup() {
+#if defined(IR_MEASURE_TIMING) && defined(IR_TIMING_TEST_PIN)
+    pinMode(IR_TIMING_TEST_PIN, OUTPUT);
+#endif
+
+    Serial.begin(115200);
+#if defined(__AVR_ATmega32U4__) || defined(SERIAL_USB) || defined(SERIAL_PORT_USBVIRTUAL)  || defined(ARDUINO_attiny3217)
+    delay(4000); // To be able to connect Serial monitor after reset or power up and before first print out. Do not wait for an attached Serial Monitor!
+#endif
+    // Just to know which program is running on my Arduino
+    Serial.println(F("START " __FILE__ " from " __DATE__ "\r\nUsing library version " VERSION_IRREMOTE));
+
+    /*
+     * Start the receiver, enable feedback LED and (if not 3. parameter specified) take LED feedback pin from the internal boards definition
+     */
+    IrReceiver.begin(IR_RECEIVE_PIN);
+    IrSender.begin(IR_SEND_PIN, ENABLE_LED_FEEDBACK); // Specify send pin and enable feedback LED at default feedback LED pin
+
+    Serial.print(F("Ready to receive IR signals at pin "));
+#if defined(ARDUINO_ARCH_STM32) || defined(ESP8266)
+    Serial.println(IR_RECEIVE_PIN_STRING);
+#else
+    Serial.println(IR_RECEIVE_PIN);
+#endif
+    Serial.print(F("Ready to send IR signals at pin "));
+#if defined(ARDUINO_ARCH_STM32) || defined(ESP8266)
+    Serial.println(IR_SEND_PIN_STRING);
+#else
+    Serial.println(IR_SEND_PIN);
+#endif
+
+#if FLASHEND > 0x1FFF && !defined(SEND_PWM_BY_TIMER) && !defined(USE_NO_SEND_PWM) && !defined(ESP32) // for esp32 we use PWM generation by ledcWrite() for each pin
+    /*
+     * Print internal signal generation info
+     */
+    IrSender.enableIROut(38);
+
+    Serial.print(F("Send signal mark duration is "));
+    Serial.print(IrSender.periodOnTimeMicros);
+    Serial.print(F(" us, pulse correction is "));
+    Serial.print((uint16_t) PULSE_CORRECTION_NANOS);
+    Serial.print(F(" ns, total period is "));
+    Serial.print(IrSender.periodTimeMicros);
+    Serial.println(F(" us"));
+
+    // infos for receive
+    Serial.print(MARK_EXCESS_MICROS);
+    Serial.println(F(" us are subtracted from all marks and added to all spaces for decoding"));
+#endif
+}
+
+uint16_t sAddress = 0x0102;
+uint8_t sCommand = 0x34;
+uint8_t sRepeats = 1;
+
+void send_ir_data() {
+    Serial.print(F("Sending: 0x"));
+    Serial.print(sAddress, HEX);
+    Serial.print(sCommand, HEX);
+    Serial.println(sRepeats, HEX);
+
+    // clip repeats at 4
+    if (sRepeats > 4) {
+        sRepeats = 4;
+    }
+    // Results for the first loop to: Protocol=NEC Address=0x102 Command=0x34 Raw-Data=0xCB340102 (32 bits)
+    IrSender.sendNEC(sAddress, sCommand, sRepeats);
+}
+
+void receive_ir_data() {
+    if (IrReceiver.decode()) {
+        Serial.print(F("Decoded protocol: "));
+        Serial.print(getProtocolString(IrReceiver.decodedIRData.protocol));
+        Serial.print(F("Decoded raw data: "));
+        Serial.print(IrReceiver.decodedIRData.decodedRawData, HEX);
+        Serial.print(F(", decoded address: "));
+        Serial.print(IrReceiver.decodedIRData.address, HEX);
+        Serial.print(F(", decoded command: "));
+        Serial.println(IrReceiver.decodedIRData.command, HEX);
+        IrReceiver.resume();
+    }
+}
+
+void loop() {
+    /*
+     * Print loop values
+     */
+    Serial.println();
+    Serial.print(F("address=0x"));
+    Serial.print(sAddress, HEX);
+    Serial.print(F(" command=0x"));
+    Serial.print(sCommand, HEX);
+    Serial.print(F(" repeats="));
+    Serial.println(sRepeats);
+    Serial.flush();
+
+    send_ir_data();
+    // wait for the receiver state machine to detect the end of a protocol
+    delay((RECORD_GAP_MICROS / 1000) + 5);
+    receive_ir_data();
+
+    // Prepare data for next loop
+    sAddress += 0x0101;
+    sCommand += 0x11;
+    sRepeats++;
+
+    delay(100); // Loop delay
+}